Clean up in an attempt to exit cleanly.
